+++ tools/tpm2_nvwrite.c
@@ -18,6 +18,7 @@ struct tpm_nvwrite_ctx {
     } auth_hierarchy;
 
     TPM2_HANDLE nv_index;
+    bool is_auth_hierarchy_specified;
 
     BYTE nv_buffer[TPM2_MAX_NV_BUFFER_SIZE];
     FILE *input_file;
@@ -25,7 +26,9 @@ struct tpm_nvwrite_ctx {
     UINT16 offset;
 };
 
-static tpm_nvwrite_ctx ctx;
+static tpm_nvwrite_ctx ctx = {
+    .is_auth_hierarchy_specified = false,
+};
 
 static tool_rc nv_write(ESYS_CONTEXT *ectx) {
 
@@ -98,26 +101,9 @@ static bool on_option(char key, char *value) {
     bool result;
     char *input_file;
     switch (key) {
-    case 'x':
-        result = tpm2_util_string_to_uint32(value, &ctx.nv_index);
-        if (!result) {
-            LOG_ERR("Could not convert NV index to number, got: \"%s\"",
-                    value);
-            return false;
-        }
-
-        if (ctx.nv_index == 0) {
-            LOG_ERR("NV Index cannot be 0");
-            return false;
-        }
-        /*
-         * If the users doesn't specify an authorization hierarchy use the index
-         * passed to -x/--index for the authorization index.
-         */
-        ctx.auth_hierarchy.ctx_path = value;
-        break;
     case 'C':
         ctx.auth_hierarchy.ctx_path = value;
+        ctx.is_auth_hierarchy_specified = true;
         break;
     case 'P':
         ctx.auth_hierarchy.auth_str = value;
@@ -152,18 +138,27 @@ static bool on_option(char key, char *value) {
     return true;
 }
 
+static bool on_arg(int argc, char **argv) {
+    /* If the users doesn't specify an authorization hierarchy use the index
+    * passed to -x/--index for the authorization index.
+    */
+    if (!ctx.is_auth_hierarchy_specified) {
+        ctx.auth_hierarchy.ctx_path = argv[0];
+    }
+    return on_arg_nv_index(argc, argv, &ctx.nv_index);
+}
+
 bool tpm2_tool_onstart(tpm2_options **opts) {
 
     const struct option topts[] = {
-        { "index",                required_argument, NULL, 'x' },
         { "hierarchy",            required_argument, NULL, 'C' },
         { "auth",                 required_argument, NULL, 'P' },
         { "input",                required_argument, NULL, 'i' },
         { "offset",               required_argument, NULL,  0  },
     };
 
-    *opts = tpm2_options_new("x:C:P:i:", ARRAY_LEN(topts), topts, on_option,
-        NULL, 0);
+    *opts = tpm2_options_new("C:P:i:", ARRAY_LEN(topts), topts, on_option,
+        on_arg, 0);
 
     return *opts != NULL;
 }